Support creating libraries with enterprise options
