+++ python/arcticdb/adapters/arctic_library_adapter.py
@@ -9,13 +9,17 @@ As of the Change Date specified in that file, in accordance with the Business So
 from abc import ABC, abstractmethod
 from typing import Iterable, List
 
-from arcticdb.options import DEFAULT_ENCODING_VERSION, LibraryOptions
+from arcticc.pb2.storage_pb2 import EnvironmentConfigsMap, LibraryConfig
+from arcticdb.config import _DEFAULT_ENV
+from arcticdb.version_store._store import NativeVersionStore
+from arcticdb.options import DEFAULT_ENCODING_VERSION, LibraryOptions, EnterpriseLibraryOptions
 from arcticc.pb2.storage_pb2 import LibraryConfig
 from arcticdb_ext.storage import Library, StorageOverride, CONFIG_LIBRARY_NAME
 from arcticdb.encoding_version import EncodingVersion
 
 
-def set_library_options(lib_desc: "LibraryConfig", options: LibraryOptions):
+def set_library_options(lib_desc: "LibraryConfig", options: LibraryOptions,
+                        enterprise_library_options: EnterpriseLibraryOptions):
     write_options = lib_desc.version.write_options
 
     write_options.dynamic_strings = True
@@ -38,6 +42,9 @@ def set_library_options(lib_desc: "LibraryConfig", options: LibraryOptions):
         options.encoding_version if options.encoding_version is not None else DEFAULT_ENCODING_VERSION
     )
 
+    write_options.sync_passive.enabled = enterprise_library_options.replication
+    write_options.delayed_deletes = enterprise_library_options.background_deletion
+
 
 class ArcticLibraryAdapter(ABC):
     @abstractmethod
@@ -58,8 +65,24 @@ class ArcticLibraryAdapter(ABC):
     def config_library(self) -> Library:
         raise NotImplementedError
 
+    def get_library_config(self, name: str, library_options: LibraryOptions,
+                           enterprise_library_options: EnterpriseLibraryOptions):
+        env_cfg = EnvironmentConfigsMap()
+
+        self.add_library_to_env(env_cfg, name)
+
+        library_options.encoding_version = (
+            library_options.encoding_version if library_options.encoding_version is not None else self._encoding_version
+        )
+        set_library_options(env_cfg.env_by_id[_DEFAULT_ENV].lib_by_path[name], library_options,
+                            enterprise_library_options)
+
+        return NativeVersionStore.create_library_config(
+            env_cfg, _DEFAULT_ENV, name, encoding_version=library_options.encoding_version
+        )
+
     @abstractmethod
-    def get_library_config(self, name: str, library_options: LibraryOptions):
+    def add_library_to_env(self, env_cfg: EnvironmentConfigsMap, name: str):
         raise NotImplementedError
 
     def get_storage_override(self) -> StorageOverride:
