+++ python/arcticdb/adapters/in_memory_library_adapter.py
@@ -47,16 +47,5 @@ class InMemoryLibraryAdapter(ArcticLibraryAdapter):
 
         return lib._library
 
-    def get_library_config(self, name, library_options: LibraryOptions):
-        env_cfg = EnvironmentConfigsMap()
-
+    def add_library_to_env(self, env_cfg: EnvironmentConfigsMap, name: str):
         add_memory_library_to_env(env_cfg, lib_name=name, env_name=_DEFAULT_ENV)
-
-        library_options.encoding_version = (
-            library_options.encoding_version if library_options.encoding_version is not None else self._encoding_version
-        )
-        set_library_options(env_cfg.env_by_id[_DEFAULT_ENV].lib_by_path[name], library_options)
-
-        return NativeVersionStore.create_library_config(
-            env_cfg, _DEFAULT_ENV, name, encoding_version=library_options.encoding_version
-        )
