+++ python/arcticdb/adapters/lmdb_library_adapter.py
@@ -147,24 +147,13 @@ class LMDBLibraryAdapter(ArcticLibraryAdapter):
 
         return lib._library
 
-    def get_library_config(self, name, library_options: LibraryOptions):
-        env_cfg = EnvironmentConfigsMap()
-
+    def add_library_to_env(self, env_cfg: EnvironmentConfigsMap, name: str):
         lmdb_config = {}
         if self._query_params.map_size:
             lmdb_config["map_size"] = self._query_params.map_size
 
         add_lmdb_library_to_env(
             env_cfg, lib_name=name, env_name=_DEFAULT_ENV, db_dir=self._path, lmdb_config=lmdb_config
-        )
-
-        library_options.encoding_version = (
-            library_options.encoding_version if library_options.encoding_version is not None else self._encoding_version
-        )
-        set_library_options(env_cfg.env_by_id[_DEFAULT_ENV].lib_by_path[name], library_options)
-
-        return NativeVersionStore.create_library_config(
-            env_cfg, _DEFAULT_ENV, name, encoding_version=library_options.encoding_version
         )
 
     def get_storage_override(self) -> StorageOverride:
