+++ python/arcticdb/adapters/mongo_library_adapter.py
@@ -63,15 +63,5 @@ class MongoLibraryAdapter(ArcticLibraryAdapter):
 
         return lib._library
 
-    def get_library_config(self, name, library_options: LibraryOptions):
-        env_cfg = EnvironmentConfigsMap()
-
+    def add_library_to_env(self, env_cfg: EnvironmentConfigsMap, name: str):
         add_mongo_library_to_env(cfg=env_cfg, lib_name=name, env_name=_DEFAULT_ENV, uri=self._uri)
-        library_options.encoding_version = (
-            library_options.encoding_version if library_options.encoding_version is not None else self._encoding_version
-        )
-        set_library_options(env_cfg.env_by_id[_DEFAULT_ENV].lib_by_path[name], library_options)
-
-        return NativeVersionStore.create_library_config(
-            env_cfg, _DEFAULT_ENV, name, encoding_version=library_options.encoding_version
-        )
