+++ python/arcticdb/version_store/library.py
@@ -10,10 +10,10 @@ import datetime
 
 import pytz
 from enum import Enum, auto
-from typing import Optional, Any, Tuple, Dict, AnyStr, Union, List, Iterable, NamedTuple
+from typing import Optional, Any, Tuple, Dict, Union, List, Iterable, NamedTuple
 from numpy import datetime64
 
-from arcticdb.options import LibraryOptions
+from arcticdb.options import LibraryOptions, EnterpriseLibraryOptions
 from arcticdb.supported_types import Timestamp
 from arcticdb.util._versions import IS_PANDAS_TWO
 
@@ -335,6 +335,7 @@ class Library:
         return self.has_symbol(symbol)
 
     def options(self) -> LibraryOptions:
+        """Library options set on this library. See also `enterprise_options`."""
         write_options = self._nvs.lib_cfg().lib_desc.version.write_options
         return LibraryOptions(
             dynamic_schema=write_options.dynamic_schema,
@@ -342,6 +343,14 @@ class Library:
             rows_per_segment=write_options.segment_row_size,
             columns_per_segment=write_options.column_group_size,
             encoding_version=self._nvs.lib_cfg().lib_desc.version.encoding_version,
+        )
+
+    def enterprise_options(self) -> EnterpriseLibraryOptions:
+        """Enterprise library options set on this library. See also `options` for non-enterprise options."""
+        write_options = self._nvs.lib_cfg().lib_desc.version.write_options
+        return EnterpriseLibraryOptions(
+            replication=write_options.sync_passive.enabled,
+            background_deletion=write_options.delayed_deletes
         )
 
     def write(
